Move moduleSuffix method from GenJVM into Symbols.
This method, previously defined in GenJVM, produces either "$" or ""
depending on certain conditions of the Symbol argument. The same logic
is needed in GenMSIL (and GenLLVM), so it was being copy/pasted.

Even for JVM backend, in Erasure, we needed to access it as
global.genJVM.moduleSuffix. Seems more reasonable to make it a method
of Symbol.

Ideally, this is a change which could be accepted upstream.