Dump classname$.sym for module classes (objects).
This represents a somewhat more sophisticated understanding of
symbols we are asked to compile in GenLLVM. I dump symbol information
to different files depending on whether it's a class or a module class
(object). Reading back in the classes works okay, although I'm still
not able to read back objects correctly.