Maybe fixes method offset bug by coalescing redundant methods.
Before we were including superfluous methods in the vtable that differ
only by return type. The pickled signature (somehow) drops or
coalesces such methods, which led to different ideas about what the
vtable contains when a file is separately compiled vs. included on the
command line.

It's not clear to me this will always work, but it does okay for the
test cases I have now.