Fix separate compilation bug related to java.lang.String.+
A line in scala.reflect.internal.Definitions adds the `+` method to
java.lang.String. In the LLVM, we dump String.sym containing this
method. On the next invocation, the compiler read the signature back
and added it again, as a duplicate. The fix is to check whether the
String class already contains `+`.

This problem doesn't occur in the JVM back-end because Java-defined
classes don't have Scala signatures (and anyway, scalac never writes
java/lang/String.class).